created projectile scene
i'd like to create a single-target damage formula that fires a projectile in a given direction. the projectile should have collision with the World, meaning if a World obstacle (eg. Tree) is between the player and the target, the projectile will hit that instead.
the projectile will originate from the player and travel in a straight line over a set distance. the angle of the line will be set via the formula target screen.
the formula target screen will need a new targeting system, which will be a straight line originating from the player, and whose origin also serves as a rotational axis point. this target method will return an angle by which we will multiply the Projectile's velocity.
i realize i should probably finish the modular formulaTargetScreen before making any new formulas, so maybe i'll attempt that first.