re-usable FormulaTargetScreen created
the 3 formulas now use the same FormulaTargetScreen scene.
it is not fully set up yet. the individual formulas still have their parameters set manually. the idea is to make a system where based on a small amount of information, the appropriate targetScreen behaviour will occur. to achieve this, a number of variables are exported:
formula_shape : Shape2D (not used yet; the shapes have been set via the inspector, for now, but in the end will be set using code)
formula_size : int (not used yet, as above; this will be used to determine the size of the targetArea collisionShape and the formulaHitbox collisionShape. i might limit the number of "sizes", for simplicity's sake. something like: TINY, SMALL, MEDIUM, LARGE, HUGE, SCREEN. though maybe i'll want to have certain formulas change size depending on other factors... whatever, we're not there yet, so won't worry for now.)
enum targetMode { NONE, FREE, ANGLE }
export(targetMode) var mode (enumerator used to determine the behaviour of inputs controlling the target screen. static AoE buffs have a target mode of NONE (0), and Flash has a target mode of FREE (1).)
export var formula_range : int (the distance from the origin point - the casting Player - that the TargetArea can travel within on FREE mode.)
export var attack_formula : bool (true or false denoting whether a formula is offensive or not. if it is, creates Targets at enemies' locations, and sets the target overlap colour to Red. if false, Targets are created at Players' locations, and sets the target_color to cyan.