formula organization
it will take creating a few more formulas of varying types to determine the best way to continue expanding the formula system. right now, the FormulaScreen is re-usable which is great.
one thing i am unsure of is that most formula calculation takes place on the formula's Hitbox script. it might be a better idea to have a generic formula Hitbox, re-usable between every formula, and simply have the formula's script itself do the calculations/assign the values. this is what i've done for the target screen: the formula node itself tells the FormulaTargetScreen the nature of the targetting behaviour, so i don't see why the formula node itself can't also tell the Hitbox the nature of the calculations - i'm talking things like potency of damage values, duration of buffs/debuffs, etc. these things take into account the stats of the player casting them. this would be worth figuring out before moving on to creating more formulas.