added Hardball formula
the formula resource is working in-game, and i started work on a new kind of targeting mode, ANGLE, which needs better controls/display.
the formula itself is not mechanically well implemented yet but the basic idea is there. next i will have to learn Raycasting and actually firing a projectile, and have that projectile able to collide with both the world and enemies, since right now it just turns on collision detection for a shape whose orientation is determined by the angle of the target_area, like other formulas.