working on Hardball formula
improving basic implementation, though it's messy.
the target screen for Hardball uses a RayShape2D to detect bodies (Targets) on its trajectory.
ISSUE: every body detected in the RayShape's path turns its Target & the reticle red; only the first Target in the RayShape's path should turn red.
pressing the directional inputs changes the angle of the target screen's KinematicBody2D (target_body), whose child is the TargetArea.
when the formula is confirmed, sets the rotation angle of the FormulaHitbox to that of the target_body:

get_parent().get_node("FormulaHitbox").rotation_degrees = target_body.rotation_degrees

...and runs the formula's start() method, which like the other formulas changes the casting player's state, deactivates their animationTree, plays the player's first Cast animation, and plays the Formula's animation.
Hardball calls 2 functions: ability_start() and ability_end(). ability_start() first shows a sprite that represents the hitbox angle (simply for visual reference, will be replaced with an actual projectile at some point - it also disappears instantly if the formula hits something).
then we check if the RayCast2D is_colliding() with anything (it looks for World objects and EnemyHurtboxes), and if it is, we check if the name of the thing it's colliding with is "Hurtbox", and if it is, we pretend the the RayCast2D is a Hitbox like any other entering the colliding body's hurtbox by manually running the collider's parent's _on_Hurtbox_area_entered function, passing $FormulaHitbox as the argument. we do this ONCE, meaning only a single Hurtbox will have its area_entered function called - the first Hurtbox the RayCast2D collides with. looking for "Hurtbox" is also useful because some things that aren't enemies have hurtboxes, like cuttable bushes & grass, so the formula will also "cut" any world objects it hits that CAN be hit.
if the collider's name ISN'T "Hurtbox", then we must have collided with something that doesn't detect hits. in this case, nothing happens.
ability_end() simply hides the sprite after a moment if it doesn't ht anything, and plays the Player's Cast_2 animation, at the end of which the formula is deleted.
as a side note, the new formula is functioning properly within the formulaData, elemental, and status systems:
Hardball levels up, and uses the same damage growth formula as Flash: potency = (base_potency*formula_level_mod) * (spell_mod*spell_mod).
Hardball's hitbox contains element = 6, which is Earth. Earth does increased damage vs. Lightning, and decreased damage vs. Earth & Wind.
Hardball's hitbox also carries the Stun status with a 50% chance of infliction. this means, against enemies with 20% status resistance, for example, Hardball has a 30% chance of Stunning them (eg. Wolf).
next is to work on the presentation of the formula. it needs sound effects for throwing it (whoosh) and for when it hits something (heavy impact+rock breaking).
it also needs a knockback vector. i need to figure out a way to translate the hitbox's angle parameter into a knockback vector to be applied to enemies hit, knocking them in the direction the projectile is moving.
i also need to refine the controls for the Target Screen. i would like, when you bring the Target Screen up, for the formula to already be aiming at the nearest enemy.