Hardball targeting aims at nearest target on cast
first, we create a Targets node as a child of World. for each enemy, we set each Target's global_position to be that of the enemy's Hurtbox, and add each Target as a child of Targets (line 78):
	target_bodies = get_tree().get_nodes_in_group(group_to_target)
	var targets = Node.new()
	targets.set_name("Targets")
	get_tree().get_root().get_node("World").add_child(targets)
	for e in range(0, target_bodies.size()):
		var target = Target.instance()
		target.global_position = target_bodies[e].hurtbox.get_child(0).global_position
		get_tree().get_root().get_node("World/Targets").add_child(target)

then, we declare a variable called closest_target, and if the size of the target_bodies array is greater than 0 (in other words, if enemies are detected), we set closest_target to first child of the hurtbox of the first index of the target_bodies array, which is the SHAPE of the hurtbox.
then we check the target_mode of the FormulaTargetScreen, which is set by the Formula. in this case, it is 2, or ANGLE. so we un-center the sprite that is to hold our reticle texture, and change the texture to be that of the white line. (line 109)
then, we check if target_bodies.size() == 0, and return out of the function if it is. in other words, if there exists no enemy to snap to, we end the function. (line 112)
otherwise, we iterate through target_bodies[], comparing the distances between each target_body to that of our Formula - which is instanced in the same location as the casting Player (line 116):

if get_parent().global_position.distance_to(target_bodies[t].global_position) < get_parent().global_position.distance_to(closest_target.global_position):

whenever we come across a distance between a target_body and the Formula that is LESS THAN the distance of the previous closest_target to the Formula, that target_body's Hurtbox shape becomes the new closest_target (line 117).
after we finish iterating through the target_bodies and comparing their distances (picking the shortest one), we have our target_body.look_at(closest_target.global_position), which sets the angle of our target_body (line) towards the nearest enemy.
in order for the above to work, i had to offset the TargetArea's CollisionShape2D (RayShape2D) by -90 degrees. this is because while KinematicBody2Ds default "facing" seems to be to the right, RayShape2Ds face downward by default. so i basically rotated the shape -90 degrees (CCW) so that it's aligned with the "facing" of the target_body, which is the thing that rotates the targeting line, and whose rotation_degrees we give to the FormulaHitbox. this is what ensures we cast in the proper direction, as confirmed by the purple line sprite when we miss.
next i'll have to come up with a way to cycle through targets in ANGLE mode.