next/previous enemy works for ANGLE target_mode
unfortunately, the order of sorting through enemies is pretty random. maybe i'll see if there's a way to sort the objects in the target_bodies array by their position.x value. that way, at least when using the next/previous buttons, it will pick the next enemy to the left/right, regardless of where they are on the y-axis. i'm not sure if this is possible, yet, though.