FormulaTargetScreen updated
changes for FREE targeting:
the maximum range of a formula now takes into account the size of its AoE. essentially, the max range of a formula should not be extended simply because it has a bigger AoE.
example:
2 formulas, A & B, have the same formula_range, but formula A has a TINY AoE and formula B has a HUGE AoE. (TINY: circle radius=8px; HUGE: circle radius=128px)
if the formula range determines how far the CENTER of the AoE can travel, then the HUGE AoE inherently has a much longer range (120px) than the TINY AoE. therefore, offsets to the formula_range need to be applied, depending on the AoE size.
the offsets are as follows:
TINY: +24 to formula_range
SMALL: +16 to formula_range
MEDIUM: no change
LARGE: -32 to formula_range
HUGE: -96 to formula_range
using the above example of formula A & B, we can see that before applying the offset, the difference of their EFFECTIVE ranges is 120px (b.radius - a.radius, or 128 - 8 = 120). if we look at the offsets applied to TINY and HUGE AoEs, we can see that they are +24 and -96, respectively. what's 24 + 96? it's 120, the difference in range, thus proving that these numbers are in fact accurate.
let's do another example to be sure, this time with SMALL and LARGE AoEs; their offsets are +16 and -32, respectively. let's call them formulas C & D.
let's assume no offset is applied, to begin with. the EFFECTIVE range of each formula is increased by the radius of its AoE. in the case of Formula C, this means 128+16=144, and in the case of Formula D, 128+64 = 192. thus the EFFECTIVE range difference is 192-144=48.
now let's look at the offsets we apply: the SMALL offset is +16, and the LARGE offset is -32. 16+32=48, which the same value as the effective range difference between formulas C & D before applying the range offset. IT JUST WORKS.
now, i've had to adjust the way next/previous works. when we scroll through targetable enemies, we check if the enemy we are scrolling to is out of range or not. the way we do this is we compare the player.global_position to the target_bodies[count].hurtbox.get_child(0).global_position, and we check if that distance exceeds the result of the following calculation:
formula_range + (target_shape.shape.radius+4)
the target_shape in this case is our targeting reticle. so we take the range of our formula and add it to the radius of the circle, then add 4. why 4? 4 is the radius of the target_bodies! so if a target_body is technically "out of range", but only slightly (ie., less than 4 pixels), the extents of its collisionShape are still technically able to overlap the AoE circle, thus it should still be targetable.
checking if out of range is simpler for the ANGLE target_mode:
player.global_position.distance_to(target_bodies[count].hurtbox.get_child(0).global_position) > formula_range + 4
it's essentially the same check only we don't factor in the radius of any AoE shapes, since the spell is simply a line, we just check if the distance to the target exceeds the length of the line + 4... +4 because the radius of the targets still exists and must be factored in. therefore, a formula with a range of 256, for example, has an effective range of 260, basically.
the FormulaTargetScreen is now one of the more mathematically complex parts of the game's code, surprisingly enough.