created custom sort function for target_bodies
the custom sort function sorts the order of enemies depending on the target_mode.
if target_mode == 1, aka FREE, then target_bodies are sorted via their position.x. in other words, when scrolling through enemies on the target screen using previous & next, next will go to the next enemy on the x-axis and so on:
func sort_target_bodies(a, b):
	if target_mode == 1 and a.position.x < b.position.x:
		return true
if target_mode == 2 aka ANGLE, then target_bodies are sorted via their angles in radians to the formula origin point, the player:
func sort_target_bodies(a, b):
	if target_mode == 2 and get_parent().global_position.angle_to_point(a.hurtbox.get_child(0).global_position) < get_parent().global_position.angle_to_point(b.hurtbox.get_child(0).global_position):
		return true

this sorts the target_bodies in a clockwise fashion around the player. in other words, pressing next will scroll to the next target clockwise, and pressing previous will scroll to the next target counterclockwise.
this was a somewhat unnecessarily complicated thing to accomplish, but i wanted to see if i could do it on my own as a challenge, which has been successful so i'm happy for now.