continuing projectile
haven't figured out the way i want it to work, yet. there are a number of reasons for complication.
the projectile needs to move at a constant speed towards a point. the projectile should disappear if it hits anything - either a World object or EnemyHurtbox - OR after it travels a certain distance (256px).
the projectile needs to get its target_point from the Hardball scene. we aim the Hardball by rotating a RayShape2D around a center axis. we need to get the global coordinates of where the RayShape2D CASTS TO, and apply that as the projectile's target.