Hardball casts projectile
got an aimed projectile working. it doesn't delete itself yet on hit, though, so for now it pierces objects & enemies. i'd like some Formulas to pierce and hit multiple enemies in a line, and some to be single-hit. i'd like this one to be single-hit so i'll figure out a way to disable it when it hits something, either a world object or an enemy hurtbox.