working on 2D physics for Projectile
there are so many ways to handle physics, i will have to experiment with them all and find the one that feels the best.
the projectile collides with world objects and deletes itself, but not when it hits enemies. this is because there are 2 separate layers for world collision and EnemyHurtbox collision. i think i need to re-do the projectile to account for both, as i think it'll be way too messy otherwise.
i will also have to experiemnt with RigidBody2D, a node i have not used yet. it seems like it might be better/more efficient to use for a projectile.
one of the issues is the the projectile has a Body and a Hitbox. if the body hits something (like a tree), it disappears, but when its Hitbox overlaps a Hurtbox, it behaves like a normal hitbox (ie., doesn't delete itself). maybe i will add a check on collision on the projectile itself to see if the thing collided with was an enemy or not, and generate/activate a hitbox on collision. now that i think about it, this might be useful for doing an angle-aimed AoE spell, like a grenade-type formula.
i think more experimentation with different nodes, their methods, and 2D physics in general is in order.