refining Projectile
Hardball working as intended. Projectile collides with World objects (areas) and Enemy bodies, and disappears on collision.
the projectile does not hit an Enemy's Hurtbox. instead, the projectile detects itself colliding with something - in this case, an Enemy BODY's CollisionShape - and then manually runs the object collided with's on_Hurtbox_area_entered(area) function, passing the data it has stored in a script attached to a child node called FormulaHitbox.
this naming convention is bad, to say the least, because it is not a Hitbox. it simply runs the operations needed to determine its damage, and stores them in variables. these variables names (formula, spell_mod, formula_level_mod, potency, knockback_vector, status, element) exist on other hitboxes. when a collision is detected from the projectile, we call the Enemy's on_Hurtbox_area_entered function manually, and pass the Node (FormulaHitbox) that stores these variables AS IF IT WERE a hitbox.
replaced the Hardball's "direction" variable with "knockback_vector", which is used to determine both the direction the enemy moves toward when hit by a projectile, and the direction of the projectile itself.