projectile Trail
working on the aesthetics for the Hardball formula. would be nice with a player baseball pitch-type throw animation. also needs some sound effects: a windy "whoosh" when flying, and a satisfying rock crumble when it impacts. also needs a visual impact effect where the rock breaks into pieces, and maybe the pieces remain on the ground.
otherwise seems to be working OK. had a lot of trouble with this one. there was a problem where ~5% of the time, we would not detect collision with an Enemy, and the projectile would simply fly right over it. i've increased the size of some of the enemy Bodies to compensate, as well as set the "safe margin" of the projectile body collision to be 8. i don't fully understand safe margin yet, but i think it has to do with collision precision: a lower number (like less than 1) is higher precision, whereas a higher number is a lower precision (8 = pixels, so begin collision calculation if within 8 pixels.. i think?).
the projectile moves at a speed of "12" (pixels per frame? i'm not sure), and its collision shape has a radius of 4. so i'm hoping a "safe margin" of 8 to compensate is enough to guarantee collision. needs more rigourous testing, but it's alright so far.