added Cleanse formula
for clearing detrimental status effects. currently just a copy of Heal that doesn't do anything yet. have to think of the best way to get rid of status effects via FormulaHitbox. also thinking that the number of status effects removed per cast could go up with the spell level.
also, weapon charge can be kept while casting. you are able to cast a formula while charging, and continue the charge when the Cast animation ends if you are holding the attack button down.