Cleanse clears a single debuff
it's implemented in a janky way. the FormulaHitbox detects the string "Cleanse" in index 0 of the formula's "status" variable. this isn't ideal because "Cleanse" isn't a status effect, per se, but whatever.
when the status "Cleanse" is detected, it simply runs a function StatusHandler.remove_debuff(body), which parses through the body's StatusDisplay and deletes the first debuff it finds before returning. this means the debuff cleansed is the one LEAST recently applied, which i might want to change to have it cleanse the debuff most recently applied. we'll see.
also doing it this way prevents more than one status from being removed. i'm not going to have the StatusHandler fuck with formula data values. the effects of the formula should be set when the formula is cast, not passed as arguments to StatusHandler.
or - i could expand the way we're currently doing it, and have one of the values determining status effects simply determine the amount of debuffs removed. like how poison & stun have durations, instead of having that value refer to a duration, it could refer to the number of statuses removed. that sounds like risky, spaghetti-code style territory, so i may want to consider creating a fresh system.