Cleanse formula working / Hardball hit effect added
Hardball sound isn't satisfying. hard to find good "rock" sounds.
Cleanse works as follows:
the Cleanse Hitbox contains a status array like any other Hitbox that causes a status - only, it is called "Cleanse", and passed as the 2nd index of the status array is the formula_level.
when the StatusHandler recieves the Cleanse status, it runs the remove_last_debuff(body, status[2]) function, where 'body' is the object from which we are to clear a debuff, and status[2] is the Cleanse formula_level.
the number of debuffs Cleansed is equal to the formula_level. we store this number in a variable, debuffs_to_cleanse.
first, we iterate BACKWARDS through the status_display node, as we want to get rid of the LAST debuff applied. we look for the variable "debuff" in each of status_display's children, starting from the most recently added. then, we check if the debuff is named "Stun", and carry on to the next child if it is. this is because Stun is not curable with Cleanse.
then, we display the damage popup of the name of the debuff being Cleansed, using the "Heal" colour. we delete the debuff, and decrement "debuffs_to_cleanse" by 1. then, we check if debuffs_to_cleanse > 0. if it is, we continue to the next loop. otherwise, we return out of the function.
i would like to Tween between the colour of the status effect normally to the Heal colour. eg. if Slow is cleansed (Slow has grey text associated with it), the Cleanse text will modulate from Grey to Cyan. the problem is the way damage_popup colours are handled, which i will need to rewrite. the name of the colour chosen should be the same name as the status associated with it. therefore, grey should not equal "Grey"; it should be "Slow". purple should not be "Purple", it should be "Poison", etc. that way, i will be able to access the colour associated with the status more easily, and apply it as the tween start point for the color modulation.