adjusted pause screen menu size
...to accommodate uniform formula display.
now some game design thoughts...
i have to consider how many formulas can be carried at once...
i think 5 is a relatively low number. but i also think that having a low number of formulas able to be carried at once incentivizes good decision-making and proper planning. however, i would like to have a large number of formulas in the game, and also want the player to feel free to experiment with as many as they want to. but, later down the line, when there are a large number of formulas implemented into the game, and lots of different TYPES and ELEMENTS of formulas, i don't want the player to have too many so as to be completely over-powered. the idea here is to "plan to be able to deal more easily with certain situations over others", not, "being able to easily deal with 100% of possible situations."
something i considered was having the spirituality stat also increase the number of carriable formulas at certain intervals. but i think that would make it too powerful. i would like formulas to "accessorize" the combat, not be the EZ solution to every combat encounter. the formulas you carry/choose to use, along with the stats you pick, will be what makes your "build". formulas will be able to be swapped at save points, so that part of the build is more malleable, but stats are more permanent.
there will also be non-combat "utility" formulas that will be used to overcome obstacles, solve puzzles, or otherwise progress through the game. things like, a light aura formula to light up otherwise pitch-black areas, or a revealer formula to reveal invisible paths over dangerous pits. or a "hookshot"-type formula, or even levitation. there could be multiple ways to traverse a single obstacle.
it's hard to say for now, but i think access to 10 formulas at once might be a good starting point. in Evermore, you could only carry 8 at once, and i thought that was too few since there were so many cool formulas in the game. by the end of secret of Mana, each spellcaster had access to 6 spirits who had 3 spells each, so 18 defense/buff spells, and 18 offensive/debuff spells. i felt like this was almost too many, at times. not because you'd get overwhelmed, but because you had so many options, you could spam spells of ANY kind and PROBABLY win. there wasn't a whole lot of thought or planning to it, just cast & win. i'd like the system for this game to be a balance of planning/"building" and freedom to experiment.