formula tuning
realized that for the formula XP application and ingredient consumption checked if the world had an instance of the formula cast with the default name and only applied XP/consumed ingredients if it was found.
instantiated formulas delete themselves when their Animation is finished. but what if the total spell cast time was less than the length of the formula's animation? then a second instance of the same formula can be cast before the first one is deleted. this didn't result in an error, but it neither consumed ingredients nor applied formula XP when this happened. it makes me realize how the formula casting system isn't as rigid as i want it to be. i'm still figuring it out/designing it as i go, but it proves to me rigourous testing is required every step of the way because unseen/unanticipated problems both can and will occur.
i might not get a lot done today other than testing & tightening the systems. i'd like everything to be intentional and error-free, without resorting to bandaid-style coding.
for example. players & enemies have hurtbox detected functions, but only the player has a "formula" hurtbox. i think i made it this way to account for both healing and formulas that didn't do damage (ie. status effects) as a separate part of the code that wouldn't take into account hit calculation. this might not be ideal. it's hard to say without more testing, which is what i'll be doing today.