cleanse/resist DamagePopup color effects; formulas extend Formula scene; EnemyStats refactor
the StatusHandler and damage_display should both take the NAMES of the statuses applied as strings to determine their display behviour, as opposed to having the damage_display need some other additional hard-coded value to determine its colour. now, the 3rd argument for display_damage_popup, which is called on a status's _ready() method, takes the name of the status as opposed to the name of the color typed as a string.
an example, in practice...
when Stun is applied to a body, its ready function calls body.hurtbox.display_damage_popup("Stunned!", false, name). before, name used to be "Yellow", and then the DamageDisplay would match "Yellow" with the proper RGB values. now, the DamageDisplay assigns colour to damagePopups based on the name of the STATUS. all that to say... if the 3rd argument of hurtbox.display_damage_popup() is "Stun", the text will display as yellow. if it is "Poison", it will display as purple. etc.
the DamagePopup now takes an additional variable, mod, which can optionally be passed as the 4th argument in hurtbox.display_damage_popup(). mod's default value is null.
mod can either be 0, 1, or null, and corresponds to how the colour of the text modulates after it's instantiated. before, we had a popup_color, and a fade(color). the fade color of most damagePopups was red; then white, if miss; and then cyan, if negative damage (heal).
when we Cleanse a status effect, we pass 1 as the 4th argument, mod. when we do this, the fade_color becomes cyan, and the popup_color becomes the color getting cleansed. so when we cleanse the Slow debuff, the damagePopup displays "Slow Cleansed!" in grey, modulates to Cyan, and fades out Cyan.
when a status effect is Resisted, we pass 0 as the 4th argument, mod. we apply similar behaviour as above, except instead of modulating to Cyan and fading to Cyan, we modulate to White and fade to White. so when Poison is resisted, for instance, it displays "Poison Resisted" in purple before modulating to white, with a white fade out.
all in all, i'm not sure about having all the different fades. i thought displaying the colour associated with the status as the word "Resisted!" works, but i'm just experimenting with ways of effectively communicating things to the player.
ALSO: formulas extend the new Formula scene, which is now the base class for all formulas. currently, it only has one method, start.
before, we used to consume ingredients and apply formula XP from the Player script: the player would instantiate a formula by pressing the formula button, then either confirm the cast or cancel it. if it was confirmed, after the "CastTimer", we checked if the Formula was there (meaning, if it hadn't been cancelled), and consumed ingredients/applied XP if it was. this was not a good way of doing things for a number of reasons, so i've redesigned it.
now, the Formulas themselves consume ingredients and apply XP. when a cast is "confirmed" from the FormulaTargetScreen, we run the Formula scene's start() method, which takes the player casting and the formula_used as arguments. we look for the ingredient cost in the formula_used's formula_reference, and remove the ingredients from the player's pouch. we also change the player's state to 9 (ACTION), deactivate the animationTree, play the Cast_1 animation, and finally play the formula's animation.
when the formula's animation ends, we apply formula XP...
player.formulaData.apply_xp_to_formula(formula_used.formula_reference.name, player.name)
...before deleting the formula.
finally, i've refactored each invidual enemy stats' script (BatStats.gd, CrowStats.gd, etc.) into one singular script called EnemyStats, which now export variables which must be set via the inspector. this was because if i wanted to add new stats to enemies, i didn't want to have to do it manually for each enemy. for example, i'm working on applying the Slow status to enemies, and need a speed_mod stat for each enemy. rather than hard-code each enemy's stat script, they all use the same one, now with editable values per enemy.