minor adjustments
increased the size of the Cleanse formula, as well as adjusted the wolf's leap range, slightly.
the speed_mod offset of the Slow status halves the base_evasion of enemies, however if the enemy itself has an evasion_mod, like the Crow's evasion_mod of 32 when it's airborne, this value is not halved and is still added to evasion to determine final hit rate.
so practical example: crow has base evasion of 16, and an evasion_mod of 32 while airborne. therefore, normally, its evasion totals 48. if it lands, we take away the evasion_mod, so it's evasion is 16 while grounded.
if slow is applied, its base_evasion of 16 is halved. therefore an airborne, slowed crow has a final evasion stat of 40 while a grounded, slowed crow has an evasion stat of 8.
i realize i should re-name evasion_mod to evasion_bonus on enemies. which i just did.