added new ingredient resources
added Oil, Mushroom, and Petal as ingredients to test with formulas. everything is almost working properly.
there was some weirdness with the FormulaUI that i have to look into. i believe the Cost dictionary ("ingredient_name": value(int)) is loaded alphabetically. meaning, for displaying ingredient costs/their icons, ingredients have to be listed in alphabetical order. the example problem i ran into was as follows:
i had changed the "Heal.tres" formula resource's Cost dictionary to
Cost = {
"Water" = 2,
"Petal" = 1
}
and Ing 1 Icon was set to the water icon, and Ing 2 Icon was set to the Petal icon.
however, in the FormulaUI, the values were mixed up, ie., next to the Petal icon was the number of Water on-hand, and vice versa. i will have to look more into detail as to why this was happening but re-ordering ingredients alphabetically seems to work.