ingredient resource issue
confirming previous issue:
the Cost dictionary of Formula resources gets sorted alphabetically, automatically. so it does not matter in which order the keys/values are input, the Cost dict keys come out alphabetically. what this means is that if the ingredients Icons are not in alphabetical order, they will not be sorted as such, since they are not loaded from a dict.
this presents an issue to me, because i want to be able to customize the formula recipe display as i want to. a secondary issue is making sure these things are put in the proper order, manually. perhaps if the Ing Icon textures were themselves put inside of a dictionary, they, too would be loaded alphabetically automatically. this trades ease of implementation for customization, though, and i prefer customization, so i will have to look into a solution.