working on expanding charge system
the charge levels/system was pretty hard-coded so i'm going to be working on making it scaleable. meaning, giving weapons their own growth/levels. weapon level will determine charge level.
right now, it doesn't work, because there is a lot of code i have to change. the charge system was originally designed a very long time ago, and is not particularly well-programmed, so i'm going to overhaul it. i am just making the preliminary changes.