preliminary charge system overhaul complete
charge system is back to working as intended. it should work more efficiently now, too.
working on removing a large amount of unnecessary variables from the player script without breaking things. variables exist like "attack_1_charged" and "attack_charging", which are checked at times when the attack button is released. for example, if you release the attack button with a charge stocked while in an evasive action, the charge attack will be queued when the evasive action finishes.
if variables like stats.charge_level are updated properly and accurately, then stats.charge_level should be the only variable checked. it has been a very long time since i worked on the Charge system, so i don't remember some of the details of why these variables are used, but i remember having trouble with charges not being reset, or lingering without the attack button being held down. or charge sound effects playing at the wrong time, or the wrong attack coming out, etc. so my solution at the time was to just keep adding variables and checks until the issues went away, as opposed to creating a rigid foundational system - toward which i am attempting to convert the current, out-of-date system.
i removed the variables "charge_count" and "charge_level_count" from the Player script. these variables were containers for the stats.charge and stats.charge_level values, and were checked at various points. now we check the stats.charge and stats.charge_level values directly.
before, stats.max_charge was a hard-coded value, 100. now, it is based off of stats.weapon_level*50 (eg. weapon_level = 2 : max_charge = 100).
when an enemy is killed with a sword attack, we get weapon_xp:

func apply_weapon_xp(who):
	weapon_xp += weapon_growth_rate
	weapon_xp_total += weapon_growth_rate
	while weapon_xp >= weapon_xp_required:
		self.weapon_level += 1
		weapon_xp -= weapon_xp_required
		weapon_xp_required *= 1.2
		weapon_xp_required = round(weapon_xp_required)
		Global.display_message_popup(who, "Sword is now level " + str(weapon_level), "level")

for now, the growth formula is identical to the formula [spell] growth formula. when we set weapon_level,  we set max_charge to weapon_level*50:

func set_weapon_level(value):
	weapon_level = value
	self.max_charge = weapon_level*50
	print("max_charge: ", max_charge)

the ChargeUI is also being overhauled, but still needs work. for now, though, it is much cleaner and the code is much clearer. the ChargeUI updates the progress bars and plays the appropriate sound effects, as it should. before, relevant charge parameters & behaviour was spread out between the Player script, the PlayerStats script, and the ChargeUI script. now, each script should only handle the relevant logic.
in summary:
the Player script checks for inputs, and runs the charge_state(_delta) function every frame while the attack button is held down in the MOVE state. after the stamina checks, it then checks the relevant charge values, and runs the chargeUI's begin_charge_1() method when a charge is commenced. the Player script is in charge of modifying the PlayerStats charge values.
the PlayerStats script is in charge of containing relevant charge data and emitting signals when that data is changed so that the ChargeUI can be updated, as well as handling weapon XP/levelling.
the ChargeUI script is in charge of audio-visually representing charge and charge_level values based on signals received from PlayerStats.