Charge overhaul working... so far
a lot of the basic logic has been changed just by a little bit, so a lot of testing will be needed to make sure things are working as intended.
so by testing, i mean:
what happens when the player...
... begins a charge - meaning, the attack input is held down at the end of an animation - on the same frame as getting hit & staggered?
... releases the attack button while charged on the same frame as getting hit & staggered?
... releases the attack button while charged during a roll/backstep, queuing a charge attack, but getting hit & staggered on the same frame the roll/backstep finishes?

the main thing to avoid is to make sure the charge gets reset properly, meaning there's no way a charge gets "locked" in accidentally, without the need for the attack input to remain pressed. there were cases where this would happen, but it was rare.

also, i've discovered a rather curious issue. i have no idea where the World Music gets unpaused following a levelUp screen. on the Player script, there is a function called "check_attack_input()" that gets called when a levelUp finishes. this was to check if the attack button was held down as the LevelUpScreen was deleted, so that if the Player was charging when they levelled up, they could continue the charge without losing it, but more importantly to guarantee a charge reset if the attack button WASN'T held down when the LevelUp finished.
anyways, there used to be an operation to unpause the World music in that player.check_attack_input() method, but i thought that was a lazy place to put it so i got rid of it. but... the music still unpauses following a level up?
there are literally 2 lines of code in the entire project that unpause the World music:
1 is in the player script for when the player recovers from "dying";
and 2 is in the World script for when the player unpauses the game.
that's it. those are literally the only 2 instances where i unpause the World music. so i have no idea what is unpausing it after a LevelUp.

other minor things: pressing ESC on the main menu quits. so you can quickly quit by pressing ESC twice. also the main manu button focus is grabbed immediately, instead of after the fade in. for faster testing purposes.